Multi session

+++ src/server.py
@@ -18,6 +18,10 @@ class Server:
         self.serverAddress = server
         self.serverRSApublic = self.serverAddress + '/serverRSApublic.pem'
         self.serverRSAprivate = self.serverAddress + '/serverRSAprivate.pem'
+        with open(self.serverRSApublic, 'rb') as f:
+            self.serverRSApublic = RSA.import_key(f.read())
+        with open(self.serverRSAprivate, 'rb') as f:
+            self.serverRSAprivate = RSA.import_key(f.read())
         self.workingDir = None
         self.currentUser = None
         self.lastMsg = 0
@@ -28,12 +32,13 @@ class Server:
             network = network.split('src')[0] + 'network'
         self.networkPath = network
         self.networkRef = None
+        self.sessions = {}
 
-    def initSession(self):
-        self.loadRSAKeys()
+    def initSession(self, resp = '', src = ''):
         self.networkRef = network_interface(self.networkPath, 'server')
         # wait for client message
-        resp = self.readMsg()
+        if resp == '' and src == '':
+            resp, src = self.readMsg()
 
         decryptRSAcipher = PKCS1_OAEP.new(self.serverRSAprivate)
         sizeOfKey = self.serverRSApublic.size_in_bytes()
@@ -48,9 +53,11 @@ class Server:
         resp = self.processResp(resp[sizeOfKey+8:])
 
         # Authenticate user
-        username, password = resp.split(":".encode("utf-8"))
+        auth_type, username, password = resp.split(":".encode("utf-8"), 3)
         h = SHA256.new(data=password)
         password = h.digest()
+        if auth_type.decode('utf-8') == "newusr":
+            self.createNewUser(username.decode('utf-8'), password)
         if (not self.authUser(username.decode('utf-8'), password)):
             print('Nice try hacker man, get outta here!')
             exit(1)
@@ -62,11 +69,19 @@ class Server:
         serverResponse = self.encMsg(self.currentUser)
         self.writeMsg(serverResponse)
 
-    def loadRSAKeys(self):
-        with open(self.serverRSApublic, 'rb') as f:
-            self.serverRSApublic = RSA.import_key(f.read())
-        with open(self.serverRSAprivate, 'rb') as f:
-            self.serverRSAprivate = RSA.import_key(f.read())
+        self.sessions[self.currentUser] = Session(self.currentUser, self.AESKey, self.msgNonce, self.workingDir, self.lastMsg, self.networkRef)
+        
+
+    def createNewUser(self, username, passHash):
+        userfolder = self.serverAddress + "/USERS/" + username
+        if os.path.exists(userfolder):
+            self.writeMsg(self.encMsg("Invalid Username"))
+        else:
+            os.mkdir(userfolder)
+            os.mkdir(userfolder + "/root")
+            with open(userfolder + "/.hash_check.hash", "wb") as f:
+                f.write(passHash)
+
 
     def authUser(self, username, passHash):
         if username in os.listdir(self.serverAddress + '/USERS'):
@@ -111,6 +126,61 @@ class Server:
     def incNonce(self):
         self.msgNonce = self.msgNonce[:8] + (int.from_bytes(self.msgNonce[8:], 'big') + 1).to_bytes(8, 'big')
 
+
+class Session:
+
+    def __init__(self, user, key, nonce, workingDir, lastMsg, networkRef, server=os.getcwd(), network=os.getcwd()):
+        if 'src' in server:
+            server = server.split('src')[0] + 'server'
+        self.serverAddress = server
+        self.workingDir = workingDir
+        self.currentUser = user
+        self.lastMsg = lastMsg
+        self.msgNonce = nonce
+        self.AESKey = key
+        # network connection
+        if 'src' in network:
+            network = network.split('src')[0] + 'network'
+        self.networkPath = network
+        self.networkRef = networkRef
+
+    def encMsg(self, message, data=b''):
+        if isinstance(message, str):
+            message = message.encode('utf-8')
+        cipher_aes = AES.new(self.AESKey, AES.MODE_GCM, self.msgNonce)
+        if(data != b''):
+            cipher_text, tag = cipher_aes.encrypt_and_digest(
+                message + " ".encode('utf-8') + data)
+        else:
+            cipher_text, tag = cipher_aes.encrypt_and_digest(message)
+        self.incNonce()
+        return tag + cipher_text
+
+    def processResp(self, resp):
+        tag = resp[:16]
+        ciphertext = resp[16:]
+        cipher_aes = AES.new(self.AESKey, AES.MODE_GCM, self.msgNonce)
+        self.incNonce()
+        try:
+            plain = cipher_aes.decrypt_and_verify(ciphertext, tag)
+            return plain
+        except ValueError:
+            print('MAC verification failed, ending session...')
+            exit(1)
+
+    def writeMsg(self, msg):
+        self.networkRef.send_msg(self.currentUser, msg)
+
+    def readMsg(self):
+        return self.networkRef.receive_msg()
+
+    def getOsPath(self):
+        return self.serverAddress + '/USERS/' + self.currentUser + self.workingDir + "/"
+
+    def incNonce(self):
+        self.msgNonce = self.msgNonce[:8] + (int.from_bytes(
+            self.msgNonce[8:], 'big') + 1).to_bytes(8, 'big')
+
     ### COMMANDS ###
 
     # • MKD – creating a folder on the server
@@ -124,41 +194,40 @@ class Server:
 
     def mkd(self, folderName):
         # makes the directory from the working directory
-        try: 
-            os.mkdir(self.getOsPath() + folderName)
-            self.writeMsg(self.encMsg("Finished"))
+        try:
+            if self.checkAddress(folderName):
+                os.mkdir(self.getOsPath() + folderName)
+                self.writeMsg(self.encMsg("Finished"))
+            else:
+                self.writeMsg(self.encMsg("Permition Failed"))
         except OSError:
             self.writeMsg(self.encMsg("Mkd Failed"))
-        
+
     def rmd(self, folderName):
         # removes a directory if it exists
         try:
-            os.rmdir(self.getOsPath() + folderName)
-            self.writeMsg(self.encMsg("Finished"))
+            if self.checkAddress(folderName):
+                os.rmdir(self.getOsPath() + folderName)
+                self.writeMsg(self.encMsg("Finished"))
+            else:
+                self.writeMsg(self.encMsg("Permition Failed"))
         except OSError:
             self.writeMsg(self.encMsg("Deletion Failed"))
-    
+
     def gwd(self):
         self.writeMsg(self.encMsg("Working directory is: " + self.workingDir))
-    
+
     def cwd(self, newDir):
-        dirs = newDir.split("/")
-        for nd in dirs:
-            if (nd == ".."):
-                if(self.workingDir == '/root'):
-                    self.writeMsg(self.encMsg(
-                        "Working directory is now: %s" % self.workingDir))
-                    return
-                else:
-                    self.workingDir = "/".join(self.workingDir.split("/")[:-1])
-            elif (nd != ".."):
-                if(os.path.exists( self.getOsPath() + newDir)):
-                    self.workingDir = self.workingDir+"/"+newDir
-        self.writeMsg(self.encMsg("Working directory is now: %s" %self.workingDir))
-    
+        if self.checkAddress(newDir):
+            self.workingDir += newDir
+        else:
+            self.writeMsg(self.encMsg("Permission Failed"))
+        self.writeMsg(self.encMsg(
+            "Working directory is now: %s" % self.workingDir))
+
     def lst(self):
         dirList = ", ".join(os.listdir(self.serverAddress + '/USERS/' +
-                                     self.currentUser + self.workingDir))
+                                       self.currentUser + self.workingDir))
         if len(dirList) == 0:
             dirList = '<empty>'
         self.writeMsg(self.encMsg(dirList))
@@ -166,7 +235,7 @@ class Server:
     def upl(self, fileName, data):
         with open(self.getOsPath()+fileName, 'wb') as f:
             f.write(data)
-        self.writeMsg(self.encMsg("%s uploaded" %fileName))
+        self.writeMsg(self.encMsg("%s uploaded" % fileName))
 
     def dnl(self, fileName):
         with open(self.getOsPath()+fileName, "rb") as f:
@@ -180,7 +249,20 @@ class Server:
         else:
             self.writeMsg(self.encMsg("The file does not exist"))
 
-
+    def checkAddress(self, addr):
+        newDir = self.workingDir
+        dirs = addr.split("/")
+        for nd in dirs:
+            if (nd == ".."):
+                if(newDir == '/root'):
+                    return False
+                else:
+                    newDir = "/".join(self.workingDir.split("/")[:-1])
+            else:
+                if(os.path.exists(self.getOsPath() + nd)):
+                    newDir = newDir+"/"+nd
+        print(newDir)
+        return True
 
 
 def main():
@@ -189,44 +271,50 @@ def main():
     s.initSession()
     while True:
         # wait for message from client
-        msg = s.readMsg()
-        msg = s.processResp(msg)
-        # parse msg into parts all msgs will be recieved iwht cmd file/foldername payload
-        msg = msg.split(' '.encode('utf-8'), 2)
-        cmd = msg[0].decode('utf-8').lower()
-        if len(msg) > 1:
-            args = msg[1:]
-            name = args[0].decode('utf-8')
-        if cmd == "mkd":
-            s.mkd(name)
-        elif cmd == "rmd":
-            s.rmd(name)
-        elif cmd == "gwd":
-            s.gwd()
-        elif cmd == "cwd":
-            s.cwd(name)
-        elif cmd == "lst":
-            s.lst()
-        elif cmd == "upl":
-            try:
-                s.upl(name, args[1])
-            except:
-                s.writeMsg(s.encMsg("Error"))
-        elif cmd == "dnl":
-            try:
-                s.dnl(name)
-            except:
-                s.writeMsg(s.encMsg("Error"))
-        elif cmd == "rmf":
-            s.rmf(name)
-        elif cmd == "END":
-            s.initSession()
+        msg, src = s.readMsg()
+        if src in s.sessions:
+            msg = s.sessions[src].processResp(msg)
+            # parse msg into parts all msgs will be recieved iwht cmd file/foldername payload
+            msg = msg.split(' '.encode('utf-8'), 2)
+            cmd = msg[0].decode('utf-8').lower()
+            if len(msg) > 1:
+                args = msg[1:]
+                name = args[0].decode('utf-8')
+            if cmd == "mkd":
+                s.sessions[src].mkd(name)
+            elif cmd == "rmd":
+                s.sessions[src].rmd(name)
+            elif cmd == "gwd":
+                s.sessions[src].gwd()
+            elif cmd == "cwd":
+                s.sessions[src].cwd(name)
+            elif cmd == "lst":
+                s.sessions[src].lst()
+            elif cmd == "upl":
+                try:
+                    s.sessions[src].upl(name, args[1])
+                except:
+                    s.sessions[src].writeMsg(s.sessions[src].encMsg("Error"))
+            elif cmd == "dnl":
+                try:
+                    s.sessions[src].dnl(name)
+                except:
+                    s.sessions[src].writeMsg(s.sessions[src].encMsg("Error"))
+            elif cmd == "rmf":
+                s.sessions[src].rmf(name)
+            elif cmd == "end_session":
+                print("bye")
+                del s.sessions[src]
+            else:
+                print(cmd)
+                s.sessions[src].writeMsg(s.sessions[src].encMsg("Invalid command"))
+            time.sleep(0.5)
+            # print client message
+            print(f"Client command: {msg}{' '*20}")
+            # time.sleep(0.5)
         else:
-            s.writeMsg(s.encMsg("Invalid command"))
-        time.sleep(0.5)
-        # print client message
-        print(f"Client command: {msg}{' '*20}")
-        time.sleep(0.5)
+            print(src)
+            s.initSession(msg, src)
 
 
 main()
